+++ src/CSS/NavBar.css
@@ -2,17 +2,25 @@
 .topnav {
     background-color: rgb(52, 58, 64);
     overflow: hidden;
+    height: 100px;
   }
   
   /* Style the links inside the navigation bar */
   a {
-    float: right;
     color: white;
-    text-align: right;
+    text-align: center;
     padding: 14px 16px;
     text-decoration: none;
     font-size: 17px;
-    border-right: 20px;
+  }
+
+  #headerConstant {
+    margin-right: 50px;
+  }
+
+  #navbarA {
+    margin-left: 10%;
+    margin-top: 80px;
   }
   
   /* Change the color of links on hover */
@@ -31,7 +39,7 @@
   }
 
   .topnav a.activeSmall{
-    float: left;
+    /* float: left; */
     background-color: rgb(52, 58, 64);
     color: white;
     border-left: 70px !important;
@@ -56,6 +64,17 @@
   padding-top: 18px;
   font-weight: bolder;
 }
+
+.logout{
+  float: right;
+  margin-right: 5%;
+}
+
+#signedInAs{
+  text-align: left;
+  margin-top: 50px;
+}
+
